requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
autopep8==1.4.4
Click==7.0
cycler==0.10.0
Cython==0.29.7
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
ImageHash==4.0
isort==4.3.18
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.3.3
Markdown==3.1
matplotlib==3.0.3
mccabe==0.6.1
mock==3.0.4
nltk==3.4.1
numpy==1.16.3
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pdf2image==1.5.4
Pillow==6.0.0
protobuf==3.7.1
pycocotools==2.0.0
pycodestyle==2.5.0
pylint==2.3.1
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
PyWavelets==1.0.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
typed-ast==1.3.5
Unidecode==1.0.23
Werkzeug==0.15.2
wrapt==1.11.1
